+++ core/src/scheduler/selector/FaissFlatPass.h
@@ -8,29 +8,44 @@
 // Unless required by applicable law or agreed to in writing, software distributed under the License
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
+#ifdef MILVUS_GPU_VERSION
+#pragma once
 
-#include "scheduler/optimizer/Optimizer.h"
+#include <condition_variable>
+#include <deque>
+#include <limits>
+#include <list>
+#include <memory>
+#include <mutex>
+#include <queue>
+#include <string>
+#include <thread>
+#include <unordered_map>
+#include <vector>
+
+#include "config/handler/GpuResourceConfigHandler.h"
+#include "scheduler/selector/Pass.h"
 
 namespace milvus {
 namespace scheduler {
 
-void
-Optimizer::Init() {
-    for (auto& pass : pass_list_) {
-        pass->Init();
-    }
-}
+class FaissFlatPass : public Pass, public server::GpuResourceConfigHandler {
+ public:
+    FaissFlatPass() = default;
 
-bool
-Optimizer::Run(const TaskPtr& task) {
-    for (auto& pass : pass_list_) {
-        if (pass->Run(task)) {
-            return true;
-        }
-    }
+ public:
+    void
+    Init() override;
 
-    return false;
-}
+    bool
+    Run(const TaskPtr& task) override;
+
+ private:
+    int64_t count_ = 0;
+};
+
+using FaissFlatPassPtr = std::shared_ptr<FaissFlatPass>;
 
 }  // namespace scheduler
 }  // namespace milvus
+#endif
