+++ core/src/scheduler/selector/FaissIVFSQ8HPass.h
@@ -24,14 +24,14 @@
 #include <vector>
 
 #include "config/handler/GpuResourceConfigHandler.h"
-#include "scheduler/optimizer/Pass.h"
+#include "scheduler/selector/Pass.h"
 
 namespace milvus {
 namespace scheduler {
 
-class FaissIVFPQPass : public Pass, public server::GpuResourceConfigHandler {
+class FaissIVFSQ8HPass : public Pass, public server::GpuResourceConfigHandler {
  public:
-    FaissIVFPQPass() = default;
+    FaissIVFSQ8HPass() = default;
 
  public:
     void
@@ -44,7 +44,7 @@ class FaissIVFPQPass : public Pass, public server::GpuResourceConfigHandler {
     int64_t count_ = 0;
 };
 
-using FaissIVFPQPassPtr = std::shared_ptr<FaissIVFPQPass>;
+using FaissIVFSQ8HPassPtr = std::shared_ptr<FaissIVFSQ8HPass>;
 
 }  // namespace scheduler
 }  // namespace milvus
