+++ core/src/scheduler/selector/FaissIVFSQ8Pass.h
@@ -24,7 +24,7 @@
 #include <vector>
 
 #include "config/handler/GpuResourceConfigHandler.h"
-#include "scheduler/optimizer/Pass.h"
+#include "scheduler/selector/Pass.h"
 
 namespace milvus {
 namespace scheduler {
