+++ core/src/scheduler/selector/Optimizer.cpp
@@ -9,29 +9,27 @@
 // is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 // or implied. See the License for the specific language governing permissions and limitations under the License.
 
-#include "scheduler/optimizer/FallbackPass.h"
-#include "scheduler/SchedInst.h"
-#include "scheduler/tasklabel/SpecResLabel.h"
+#include "scheduler/selector/Optimizer.h"
 
 namespace milvus {
 namespace scheduler {
 
 void
-FallbackPass::Init() {
+Optimizer::Init() {
+    for (auto& pass : pass_list_) {
+        pass->Init();
+    }
 }
 
 bool
-FallbackPass::Run(const TaskPtr& task) {
-    auto task_type = task->Type();
-    if (task_type != TaskType::SearchTask && task_type != TaskType::BuildIndexTask) {
-        return false;
+Optimizer::Run(const TaskPtr& task) {
+    for (auto& pass : pass_list_) {
+        if (pass->Run(task)) {
+            return true;
+        }
     }
-    // NEVER be empty
-    LOG_SERVER_DEBUG_ << "FallbackPass!";
-    auto cpu = ResMgrInst::GetInstance()->GetCpuResources()[0];
-    auto label = std::make_shared<SpecResLabel>(cpu);
-    task->label() = label;
-    return true;
+
+    return false;
 }
 
 }  // namespace scheduler
