+++ core/unittest/scheduler/CMakeLists.txt
@@ -21,7 +21,7 @@ set(test_files
         ${CMAKE_CURRENT_SOURCE_DIR}/test_scheduler.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/test_task.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/test_job.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/test_optimizer.cpp
+        ${CMAKE_CURRENT_SOURCE_DIR}/test_selector.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/test_tasktable.cpp)
 
 add_executable(test_scheduler
